making abstraction more go-idiomatic
