+++ queue/publisher/options.go
@@ -1,7 +1,7 @@
 package publisher
 
 import (
-	"github.com/Azure/go-shuttle/common/baseinterfaces"
+	"github.com/Azure/go-shuttle/common"
 	"github.com/Azure/go-shuttle/common/options/publisheropts"
 	"time"
 
@@ -49,7 +49,7 @@ func SetDefaultHeader(headerName, msgKey string) ManagementOption {
 // SetDuplicateDetection guarantees that the topic will have exactly-once delivery over a user-defined span of time.
 // Defaults to 30 seconds with a maximum of 7 days
 func WithDuplicateDetection(window *time.Duration) ManagementOption {
-	return func(p baseinterfaces.BasePublisher) error {
+	return func(p common.Publisher) error {
 		p.(QueuePublisher).AppendQueueManagementOption(servicebus.QueueEntityWithDuplicateDetection(window))
 		return nil
 	}
